+++ apps/nextjs-website/src/components/templates/TutorialTemplate/TutorialTemplate.tsx
@@ -4,7 +4,7 @@ import { FragmentProvider } from '@/components/organisms/FragmentProvider/Fragme
 import { Box, Typography } from '@mui/material';
 import RelatedLinks from '@/components/atoms/RelatedLinks/RelatedLinks';
 import ProductLayout from '@/components/organisms/ProductLayout/ProductLayout';
-import { StrapiTutorials } from '@/lib/strapi/tutorial';
+import { StrapiTutorials } from '@/lib/strapi/fetchTutorials';
 import { Product } from '@/lib/types/product';
 import PartRenderer from '@/components/molecules/PartRenderer/PartRenderer';
 import { Part } from '@/lib/types/part';
