+++ apps/nextjs-website/src/lib/strapi/__tests__/fixtures/media.ts
@@ -1,0 +1,22 @@
+export const mediaJson = {
+  data: {
+    id: 4,
+    attributes: {
+      name: 'webinar-cover-io-remote-content.jpg',
+      alternativeText: null,
+      caption: null,
+      width: 728,
+      height: 416,
+      hash: 'webinar_cover_io_remote_content_62f1f615b5',
+      ext: '.jpg',
+      mime: 'image/jpeg',
+      size: 30.06,
+      url: 'http://localhost:1337/uploads/webinar_cover_io_remote_content_62f1f615b5.jpg',
+      previewUrl: null,
+      provider: 'strapi-provider-upload-custom',
+      provider_metadata: null,
+      createdAt: '2024-04-15T14:25:47.773Z',
+      updatedAt: '2024-08-19T13:39:34.462Z',
+    },
+  },
+};
