+++ apps/nextjs-website/src/lib/strapi/fetches/fetchTutorialListPages.ts
@@ -1,0 +1,22 @@
+import * as qs from 'qs';
+import { fetchFromStrapi } from '../fetchFromStrapi';
+import { StrapiTutorialListPagesCodec } from '../codecs/TutorialListPagesCodec';
+
+const makeStrapiTutorialListPagePopulate = () =>
+  qs.stringify({
+    populate: {
+      product: '*',
+      tutorials: {
+        populate: ['image'],
+      },
+      bannerLinks: {
+        populate: ['icon'],
+      },
+    },
+  });
+
+export const fetchTutorialListPages = fetchFromStrapi(
+  'tutorial-list-pages',
+  makeStrapiTutorialListPagePopulate(),
+  StrapiTutorialListPagesCodec
+);
