+++ apps/nextjs-website/src/lib/tutorials.ts
@@ -1,5 +1,5 @@
 import { partFromStrapiPart } from './strapi/codecs/PartCodec';
-import { StrapiTutorials } from './strapi/tutorial';
+import { StrapiTutorials } from './strapi/fetches/fetchTutorials';
 import { Part } from './types/part';
 import { Tutorial } from './types/tutorialData';
